Get rid of figures no longer needed
