+++ pf2/figures/figureS3.py
@@ -1,22 +1,24 @@
 """Figure S3"""
 
-import anndata
-from .common import getSetup
-from .commonFuncs.plotFactors import plot_gene_factors_partial
+from anndata import read_h5ad
+from .common import getSetup, subplotLabel
+from .commonFuncs.plotPaCMAP import plot_wp_pacmap
+from ..data_import import add_obs
+
 
 
 def makeFigure():
-    """Get a list of the axis objects and create a figure."""
-    ax, f = getSetup((12, 8), (5, 6))
+    ax, f = getSetup((12, 12), (4, 4))
+    subplotLabel(ax)
 
-    X = anndata.read_h5ad("/opt/northwest_bal/full_fitted.h5ad")
+    X = read_h5ad("/opt/northwest_bal/full_fitted.h5ad", backed="r")
+    add_obs(X, "patient_category")
+    X = X[X.obs["patient_category"] != "Non-Pneumonia Control"] 
 
     for i, cmp in enumerate([3, 10, 14, 15, 16, 23, 34, 55, 67, 22, 62, 1, 4]):
-        plot_gene_factors_partial(cmp, X, ax[2 * i], geneAmount=10, top=True)
-        plot_gene_factors_partial(cmp, X, ax[2 * i + 1], geneAmount=10, top=False)
+        plot_wp_pacmap(X, cmp, ax[i], cbarMax=0.4)
         
-        
-    for i in [26, 27, 28, 29]:
+    for i in [13, 14, 15]:
         ax[i].remove()
